Remove jsapi.hpp and feed .cpp into bindgen
